+++ functional-tests/functional/input/src/cpp/AttributesImpl.h
@@ -35,8 +35,6 @@ public:
     void set_struct_attribute( const ::test::Attributes::ExampleStruct& value ) override;
     ::test::Attributes::StringArray get_array_attribute( ) const override;
     void set_array_attribute( const ::test::Attributes::StringArray& value ) override;
-    std::string get_internal_setter_attribute( ) const override;
-    void set_internal_setter_attribute( const std::string& value ) override;
 
 private:
     uint32_t m_build_in_type_attribute = 0;